vector graphic rendering for different zoom levels

+++ source/ModelBasic/SimulationAccess.h
@@ -18,7 +18,7 @@ public:
 	virtual void requireData(IntRect rect, ResolveDescription const& resolveDesc) = 0;
     virtual void requireData(ResolveDescription const& resolveDesc) = 0;
     virtual void requirePixelImage(IntRect rect, QImagePtr const& target, std::mutex& mutex) = 0;
-    virtual void requireVectorImage(IntRect rect, int zoom, QImagePtr const& target, std::mutex& mutex) = 0;
+    virtual void requireVectorImage(IntRect rect, double zoom, QImagePtr const& target, std::mutex& mutex) = 0;
     virtual void selectEntities(IntVector2D const& pos) = 0;
     virtual void deselectAll() = 0;
     virtual void applyAction(PhysicalAction const& action) = 0;
